Refactored Profile to use tab navigation, added images to party views
